# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Laravel API",
        "description": "這是我們的 API 文件",
        "contact": {
            "email": "xxx"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/createdept": {
            "post": {
                "tags": [
                    "base_dept"
                ],
                "summary": "新增部門資訊",
                "description": "新增部門資訊",
                "operationId": "createdept",
                "parameters": [
                    {
                        "name": "dept_no",
                        "in": "query",
                        "description": "部門代號",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "dept_nm",
                        "in": "query",
                        "description": "部門名稱",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "note",
                        "in": "query",
                        "description": "備註",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_valid",
                        "in": "query",
                        "description": "是否有效",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "uuid": {
                                            "type": "string",
                                            "example": "0b422f02-5acf-4bbb-bddf-4f6fdd843b08"
                                        },
                                        "dept_no": {
                                            "type": "string",
                                            "example": "A02"
                                        },
                                        "dept_nm": {
                                            "type": "string",
                                            "example": "財務處"
                                        },
                                        "note": {
                                            "type": "string",
                                            "example": "測試測試"
                                        },
                                        "is_valid": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "create_user": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "create_time": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "update_user": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001975Z"
                                        },
                                        "update_time": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001986Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "未找到部門"
                    }
                }
            }
        },
        "/api/dept/{deptno}": {
            "get": {
                "tags": [
                    "base_dept"
                ],
                "summary": "查詢特定部門資訊",
                "description": "查詢特定部門資訊",
                "operationId": "getdeptno",
                "parameters": [
                    {
                        "name": "deptno",
                        "in": "path",
                        "description": "部門代號",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "uuid": {
                                            "type": "string",
                                            "example": "0b422f02-5acf-4bbb-bddf-4f6fdd843b08"
                                        },
                                        "dept_no": {
                                            "type": "string",
                                            "example": "A02"
                                        },
                                        "dept_nm": {
                                            "type": "string",
                                            "example": "財務處"
                                        },
                                        "note": {
                                            "type": "string",
                                            "example": "測試測試"
                                        },
                                        "is_valid": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "create_user": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "create_time": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "update_user": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001975Z"
                                        },
                                        "update_time": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001986Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "未找到部門"
                    }
                }
            }
        },
        "/api/dept2/{dept_nm}": {
            "get": {
                "tags": [
                    "base_dept"
                ],
                "summary": "查詢特定部門資訊",
                "description": "查詢特定部門資訊",
                "operationId": "getdeptnm",
                "parameters": [
                    {
                        "name": "dept_nm",
                        "in": "path",
                        "description": "部門名稱",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "uuid": {
                                            "type": "string",
                                            "example": "0b422f02-5acf-4bbb-bddf-4f6fdd843b08"
                                        },
                                        "dept_no": {
                                            "type": "string",
                                            "example": "A02"
                                        },
                                        "dept_nm": {
                                            "type": "string",
                                            "example": "財務處"
                                        },
                                        "note": {
                                            "type": "string",
                                            "example": "測試測試"
                                        },
                                        "is_valid": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "create_user": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "create_time": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "update_user": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001975Z"
                                        },
                                        "update_time": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001986Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "未找到部門"
                    }
                }
            }
        },
        "/api/depts/valid": {
            "get": {
                "tags": [
                    "base_dept"
                ],
                "summary": "查詢所有有效部門資訊",
                "description": "查詢所有有效部門資訊",
                "operationId": "getalldept",
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "uuid": {
                                            "type": "string",
                                            "example": "0b422f02-5acf-4bbb-bddf-4f6fdd843b08"
                                        },
                                        "dept_no": {
                                            "type": "string",
                                            "example": "A02"
                                        },
                                        "dept_nm": {
                                            "type": "string",
                                            "example": "財務處"
                                        },
                                        "note": {
                                            "type": "string",
                                            "example": "測試測試"
                                        },
                                        "is_valid": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "create_user": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "create_time": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "update_user": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001975Z"
                                        },
                                        "update_time": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001986Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "未找到有效部門"
                    }
                }
            }
        },
        "/api/dept/{deptno}/disable": {
            "patch": {
                "tags": [
                    "base_dept"
                ],
                "summary": "刪除特定部門資訊",
                "description": "刪除特定部門資訊",
                "operationId": "deletedept",
                "parameters": [
                    {
                        "name": "deptno",
                        "in": "path",
                        "description": "部門代號",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "uuid": {
                                            "type": "string",
                                            "example": "0b422f02-5acf-4bbb-bddf-4f6fdd843b08"
                                        },
                                        "dept_no": {
                                            "type": "string",
                                            "example": "A02"
                                        },
                                        "dept_nm": {
                                            "type": "string",
                                            "example": "財務處"
                                        },
                                        "note": {
                                            "type": "string",
                                            "example": "測試測試"
                                        },
                                        "is_valid": {
                                            "type": "string",
                                            "example": "0"
                                        },
                                        "create_user": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "create_time": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "update_user": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001975Z"
                                        },
                                        "update_time": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001986Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "未找到部門"
                    }
                }
            }
        },
...
    },
...
}